+++ Tests/EntityFrameworkCore.MySql/EntityFrameworkCore.MySql.Infrastructure/Persistence/Configurations/Associations/AuthorConfiguration.cs
@@ -16,6 +16,13 @@ namespace EntityFrameworkCore.MySql.Infrastructure.Persistence.Configurations.As
 
             builder.Property(x => x.Name)
                 .IsRequired();
+
+            builder.Property(x => x.Surname)
+                .IsRequired();
+
+            builder.Property(x => x.DisplayName)
+                .IsRequired()
+                .HasComputedColumnSql(@"[Name], [Surname]");
         }
     }
 }